Safari-609.1.11.1/LayoutTests/platform/ios/webgl/2.0.0/conformance/context/context-hidden-alpha-expected.txt

CONSOLE MESSAGE: line 154: FAIL gl.getParameter(gl.ALPHA_BITS) should be 0. Was 8.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance/context/context-hidden-alpha.html
[ 1: PASS ] [object WebGLRenderingContext] is non-null.
[ 2: FAIL ] gl.getParameter(gl.ALPHA_BITS) should be 0. Was 8.
[ 3: PASS ] [object WebGLProgram] is non-null.
[ 4: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 5: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
[ 6: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 7: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
[ 8: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 9: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 10: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
[ 11: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 12: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAEaElEQVR4Ae3UAQkAMAwDwW3K6l9VmYyHq4JwKbkzcxwBAgQKAq8QUkYCBAh8AYPlDwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEDJYfIEAgI2CwMlUJSoCAwfIDBAhkBAxWpipBCRAwWH6AAIGMgMHKVCUoAQIGyw8QIJARMFiZqgQlQMBg+QECBDICBitTlaAECBgsP0CAQEbAYGWqEpQAAYPlBwgQyAgYrExVghIgYLD8AAECGQGDlalKUAIEFhKbAeyGENqrAAAARGVYSWZNTQAqAAAACAABh2kABAAAAAEAAAAaAAAAAAADoAEAAwAAAAEAAQAAoAIABAAAAAEAAAEsoAMABAAAAAEAAACWAAAAAElLQfcAAAAASUVORK5CYII=. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAABLKADAAQAAAABAAAAlgAAAABJS0H3AAAFOUlEQVR4Ae3dsW1kQRBDwdPhAtvQN7SDAhjar4GSOe0QxQ9iPf18Pp8//ggQIHBB4O+FkDISIEDgV8Bg+Q4IEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRAwWL4BAgTOCBisM1UJSoCAwfINECBwRsBgnalKUAIEDJZvgACBMwIG60xVghIgYLB8AwQInBEwWGeqEpQAAYPlGyBA4IyAwTpTlaAECBgs3wABAmcEDNaZqgQlQMBg+QYIEDgjYLDOVCUoAQIGyzdAgMAZAYN1pipBCRD4d5Hg+/1ejC0zgZTAxf+Y5RdW6hMShgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASsBgpeoQhgCBJWCwlo4bAQIpAYOVqkMYAgSWgMFaOm4ECKQEDFaqDmEIEFgCBmvpuBEgkBIwWKk6hCFAYAkYrKXjRoBASuDn8/mkAglDgACBl4BfWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgIG6yXjnQCBnIDBylUiEAECLwGD9ZLxToBATsBg5SoRiACBl4DBesl4J0AgJ2CwcpUIRIDAS8BgvWS8EyCQEzBYuUoEIkDgJWCwXjLeCRDICRisXCUCESDwEjBYLxnvBAjkBAxWrhKBCBB4CRisl4x3AgRyAgYrV4lABAi8BAzWS8Y7AQI5AYOVq0QgAgReAv8BmboFrXdiuqMAAABEZVhJZk1NACoAAAAIAAGHaQAEAAAAAQAAABoAAAAAAAOgAQADAAAAAQABAACgAgAEAAAAAQAAASygAwAEAAAAAQAAAJYAAAAASUtB9wAAAABJRU5ErkJggg==.
[ 13: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 14: PASS ] successfullyParsed is true
[ FAIL ] 9 failures reported
